moved `utility::debug` to `debug_utils.hpp/cpp`
